AzureFunctions.Python/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.3.0
azure-storage-blob==12.4.0
python-dateutil==2.8.1
jsonpickle==1.4.1
